+++ test/framework/src/main/java/org/opensearch/index/engine/EngineTestCase.java
@@ -112,6 +112,8 @@ import org.opensearch.index.seqno.LocalCheckpointTracker;
 import org.opensearch.index.seqno.ReplicationTracker;
 import org.opensearch.index.seqno.RetentionLeases;
 import org.opensearch.index.seqno.SequenceNumbers;
+import org.opensearch.index.shard.ShardPath;
+import org.opensearch.index.store.FsDirectoryFactory;
 import org.opensearch.index.store.Store;
 import org.opensearch.index.translog.InternalTranslogManager;
 import org.opensearch.index.translog.LocalTranslog;
@@ -364,6 +366,12 @@ public abstract class EngineTestCase extends OpenSearchTestCase {
         }
     }
 
+    protected static ParseContext.Document testContextSpecificDocument() {
+        ParseContext.Document doc = testDocumentWithTextField("criteria");
+        doc.setGroupingCriteria("grouping_criteria");
+        return doc;
+    }
+
     protected static ParseContext.Document testDocumentWithTextField() {
         return testDocumentWithTextField("test");
     }
@@ -525,7 +533,17 @@ public abstract class EngineTestCase extends OpenSearchTestCase {
     }
 
     protected Store createStore(final IndexSettings indexSettings, final Directory directory) throws IOException {
-        return new Store(shardId, indexSettings, directory, new DummyShardLock(shardId));
+        final Path path = createTempDir().resolve(shardId.getIndex().getUUID()).resolve(String.valueOf(shardId.id()));
+        final ShardPath shardPath = new ShardPath(false, path, path, shardId);
+        return new Store(
+            shardId,
+            indexSettings,
+            directory,
+            new DummyShardLock(shardId),
+            Store.OnClose.EMPTY,
+            shardPath,
+            new FsDirectoryFactory()
+        );
     }
 
     protected Translog createTranslog(LongSupplier primaryTermSupplier) throws IOException {